added sample implementation
git-svn-id: https://btanks.svn.sourceforge.net/svnroot/btanks/trunk@6460 5c0bf991-d514-0410-a65f-ce4832e571f3

+++ clunk/context.cpp
@@ -6,7 +6,7 @@ Context::Context() : sample_rate(0), channels(0), period_size(0) {
 
 }
 
-void Context::init(const int sample_rate, const int channels, int period_size) {
+void Context::init(const int sample_rate, const Uint8 channels, int period_size) {
 
 }
