+++ clunk/context.h
@@ -2,18 +2,19 @@
 #define CLUNK_CONTEXT_H__
 
 #include "export_clunk.h"
+#include <SDL_audio.h>
 
 namespace clunk {
 class CLUNKAPI Context {
 public: 
 	Context();
 	
-	void init(const int sample_rate, const int channels, int period_size);
+	void init(const int sample_rate, const Uint8 channels, int period_size);
 	void deinit();
 	
 	~Context();
 private: 
-	int sample_rate, channels, period_size;
+	Uint8 sample_rate, channels, period_size;
 };
 }
